+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Add support for ActiveRecord 8.1
 - remove BabySqueel::Nodes::Node and Arel::Math
 - remove Operators::Equality, Operators::ArelAliasing
 - remove functions